Refactored ui and energy storage
stored energy is now a float in the player
used vertical and horizontal boxes for scaling health and energy bars on player ui

+++ CardslingerPrototype/Source/CardslingerPrototype/BaseCharacterClass.cpp
@@ -366,5 +366,5 @@ float ABaseCharacterClass::GetHealthPercent() const
 
 float ABaseCharacterClass::GetEnergyPercent() const
 {
-	return (float)CurrentEnergy/(float)MaxEnergy;
+	return CurrentEnergy/MaxEnergy;
 }
